Implement Repository pattern to refactor functionalities
