+++ app/Http/Controllers/CookbookController.php
@@ -2,9 +2,9 @@
 
 namespace App\Http\Controllers;
 
-use App\Cookbook;
 use Tymon\JWTAuth\JWTAuth;
 use Illuminate\Http\Request;
+use App\Http\Repositories\CookbookRepository;
 
 /**
  * Class UserController
@@ -13,25 +13,21 @@ use Illuminate\Http\Request;
  */
 class CookbookController extends Controller
 {
+    protected $cookbook;
+
     /**
      * Constructor
      *
-     * @param JWTAuth $jwt jwt
+     * @param JWTAuth            $jwt      jwt
+     * @param CookbookRepository $cookbook cookbookRepository
+     *
+     * @throws \Tymon\JWTAuth\Exceptions\JWTException
      */
-    public function __construct(JWTAuth $jwt)
+    public function __construct(JWTAuth $jwt, CookbookRepository $cookbook)
     {
-        $this->middleware('jwt.auth', ['only' => ['update', 'store', 'destroy']]);
         $this->jwt = $jwt;
-
         $this->user = $this->jwt->parseToken()->authenticate();
-
-        if (! $this->user ) {
-            return response()->json(
-                [
-                    'msg' => 'user not authenticated'
-                ]
-            );
-        }
+        $this->cookbook = $cookbook;
     }
 
     /**
@@ -41,32 +37,7 @@ class CookbookController extends Controller
      */
     public function index()
     {
-        $cookbooks = Cookbook::with('Recipes', 'Users')
-            ->where('user_id', $this->jwt->toUser()->id)
-            ->get();
-
-        $data = $cookbooks->toArray();
-
-        $meta = [];
-
-        foreach ($data as $key => $val) {
-            $meta[$key] = $val;
-            $meta[$key]['links'] = [
-                'methods' => [
-                    'get' => 'api/v1/cookbook/' . $val["id"],
-                    'put' => 'api/v1/cookbook/' . $val["id"],
-                    'delete' => 'api/v1/cookbook/' . $val["id"]
-                ]
-            ];
-        }
-
-        return response()->json(
-            [
-                'response' => [
-                    'cookbooks' => $meta
-                ]
-            ], 200
-        );
+        return $this->cookbook->index($this->jwt);
     }
 
     /**
@@ -78,8 +49,6 @@ class CookbookController extends Controller
      */
     public function store(Request $request)
     {
-        $response = [];
-
         $this->validate(
             $request, [
                 'name' => 'required',
@@ -87,78 +56,20 @@ class CookbookController extends Controller
             ]
         );
 
-        $cookbook = new Cookbook();
-
-        $cookbook->name = $request->input('name');
-        $cookbook->description = $request->input('description');
-        $cookbook->user_id = $this->user->id;
-
-        try {
-            if ($cookbook->save()) {
-                $response['created'] = true;
-                $response['status'] = 200;
-                $response['links'] = [
-                    'get' => 'api/v1/cookbooks/' . $cookbook->id,
-                    'put' => 'api/v1/cookbooks/' . $cookbook->id,
-                    'patch' => 'api/v1/cookbooks/' . $cookbook->id,
-                    'delete' => 'api/v1/cookbooks/' . $cookbook->id
-                ];
-            }
-        } catch (Exception $e) {
-            $response["error"] = $e->getMessage();
-            $response["status"] = 422;
-        }
-
-
-        return response()->json(
-            [
-                'response' => $response
-            ], $response["status"]
-        );
+        return $this->cookbook->store($request, $this->user);
     }
 
     /**
      * Update cookbook
      *
-     * @param Request $request    request input
+     * @param Request $request    req
      * @param int     $cookbookId paramname
      *
      * @return \Illuminate\Http\JsonResponse
      */
     public function update(Request $request, $cookbookId)
     {
-        $response = [];
-
-        $cookbook = self::cookbookExist($cookbookId);
-
-        if (! $cookbook || $cookbook === null) {
-            $response["error"] = 'Cookbook does not exist.';
-            $response["status"] = 404;
-        } else {
-            $fields = $request->only('name', 'description');
-
-            foreach ($fields as $key => $val) {
-                if ($val !== null || !is_null($val)) {
-                    $cookbook->$key = $val;
-                }
-            }
-
-            try {
-                if ($cookbook->save()) {
-                    $response["updated"] = true;
-                    $response["status"] = 204;
-                }
-            } catch (Exception $e) {
-                $response["error"] = $e->getMessage();
-                $response["status"] = 422;
-            }
-        }
-
-        return response()->json(
-            [
-                'response' => $response
-            ], $response["status"]
-        );
+        return $this->cookbook->update($request, $cookbookId);
     }
 
     /**
@@ -170,41 +81,6 @@ class CookbookController extends Controller
      */
     public function delete($cookbookId)
     {
-        $response = [];
-
-        $cookbook = self::cookbookExist($cookbookId);
-
-        if (! $cookbook || $cookbook === null) {
-            $response["error"] = 'Cookbook does not exist.';
-            $response["status"] = 404;
-        } else {
-            try {
-                if ($cookbook->delete()) {
-                    $response["deleted"] = true;
-                    $response["status"] = 202;
-                }
-            } catch (Exception $e) {
-                $response["error"] = $e->getMessage();
-                $response["status"] = 422;
-            }
-        }
-
-        return response()->json(
-            [
-                'response' => $response
-            ], $response["status"]
-        );
-    }
-
-    /**
-     * Find the cookbook
-     *
-     * @param int $cookbookId cookbokId
-     *
-     * @return mixed
-     */
-    public static function cookbookExist($cookbookId)
-    {
-        return Cookbook::findOrFail($cookbookId);
+        return $this->cookbook->delete($cookbookId);
     }
 }
