+++ app/Http/Controllers/RecipeController.php
@@ -2,9 +2,9 @@
 
 namespace App\Http\Controllers;
 
-use App\Recipe;
 use Tymon\JWTAuth\JWTAuth;
 use Illuminate\Http\Request;
+use App\Http\Repositories\RecipeRepository;
 
 /**
  * Class UserController
@@ -13,15 +13,21 @@ use Illuminate\Http\Request;
  */
 class RecipeController extends Controller
 {
+    protected $recipe;
+
     /**
      * Constructor
      *
-     * @param JWTAuth $jwt auth-jwt
+     * @param JWTAuth          $jwt    auth-jwt
+     * @param RecipeRepository $recipe recipeRepository
+     *
+     * @throws \Tymon\JWTAuth\Exceptions\JWTException
      */
-    public function __construct(JWTAuth $jwt)
+    public function __construct(JWTAuth $jwt, RecipeRepository $recipe)
     {
         $this->jwt = $jwt;
         $this->user = $this->jwt->parseToken()->authenticate();
+        $this->recipe = $recipe;
     }
 
     /**
@@ -31,25 +37,17 @@ class RecipeController extends Controller
      */
     public function index()
     {
-        return response(
-            [
-                'data' =>  Recipe::with('Cookbook')
-                    ->where('user_id', $this->jwt->toUser()->id)
-                    ->get()
-                    ->toArray()
-            ]
-        );
+        return $this->recipe->index($this->jwt);
     }
 
     /**
      * Create recipe for user
      *
-     * @param Request            $request  Form input
-     * @param CookbookController $cookbook cookbook
+     * @param Request $request Form input
      *
      * @return \Illuminate\Http\JsonResponse
      */
-    public function store(Request $request, CookbookController $cookbook)
+    public function store(Request $request)
     {
         $this->validate(
             $request, [
@@ -61,35 +59,7 @@ class RecipeController extends Controller
             ]
         );
 
-        $recipe = new Recipe(
-            [
-                'name' => $request->name,
-                'description' => $request->description,
-                'imgUrl' => $request->url,
-                'ingredients' => $request->ingredients,
-                'user_id' => $this->user->id
-            ]
-        );
-
-        try {
-            if ($cookbook::cookbookExist($request->cookbookId)) {
-                $recipe->cookbook_id = $request->cookbookId;
-                $recipe->save();
-
-                $data = $recipe;
-                $statusCode = 201;
-            }
-        } catch(\Exception $e){
-            $data = null;
-            $statusCode = 404;
-        }
-
-        return response(
-            [
-                "data" => $data,
-                'status' => $data ? 'success' : 'error or unknown cookbook.'
-            ], $statusCode ?? 200
-        );
+        return $this->recipe->store($request, $this->user);
     }
 
     /**
@@ -102,20 +72,7 @@ class RecipeController extends Controller
      */
     public function update(Request $request, $recipeId)
     {
-        try {
-            $recipe = self::recipeExist($recipeId);
-            $recipe->update($request->all());
-        } catch(\Exception $e) {
-            $recipe = null;
-            $statusCode = 404;
-        }
-
-        return response(
-            [
-                "data" => $recipe,
-                "updated" => $recipe ? true : "error",
-            ], $statusCode ?? 204
-        );
+        return $this->recipe->update($request, $recipeId);
     }
 
     /**
@@ -127,32 +84,6 @@ class RecipeController extends Controller
      */
     public function delete($recipeId)
     {
-        try {
-            $recipe = self::recipeExist($recipeId);
-            $deleted = $recipe->delete();
-            $statusCode = $deleted ? 202 : 422;
-        } catch (\Exception $e) {
-            $deleted = false;
-            $statusCode = 404;
-        }
-
-        return response(
-            [
-                'deleted' => $deleted,
-                'status' => $deleted ? "success" : "error",
-            ], $statusCode
-        );
-    }
-
-    /**
-     * Find the cookbook
-     *
-     * @param int $recipeId $recipeId
-     *
-     * @return mixed
-     */
-    protected static function recipeExist($recipeId)
-    {
-        return Recipe::findorFail($recipeId);
+        return $this->recipe->delete($recipeId);
     }
 }
